Revert backend format to iCal / Add Lobject to backend
It is more useful to keep iCal as the backend format because it is the key component of the application.
Moreover, we can translate iCal to jCal easily.

However, we needed a format where we could process our search request. So, we are now using the Lobject format too. It as just a Linear json Object with the fields we use.